refactored validate function with documentation
